feat(tools): distill_lab — diagnostic harness for the sleep pipeline
Workbench (not CI) calling the PRODUCTION distill/hub/weave functions so
diagnostics never drift from what the kernel runs: 'pipeline' simulates
gen-0/1/2 over a text, 'probe' runs the golden mini-set, 'engram' does a
hot before/after quality test on a live engram (dry-run default).
'bakeoff' (model audition) reserved for a design round with the Operator.
Also: mypy no-any-return fix in axons.py, unused import in strip_axons.py
(both flagged live by the Sentinel).

+++ src/red_pill/metabolism/axons.py
@@ -61,7 +61,7 @@ def compute_axon_weight(similarity: float, delta_seconds: float) -> float:
 	"""W = α·sim + (1-α)·(1 - Δt/Δt_max). Callers gate on cfg.AXON_GATE."""
 	dt_max = cfg.AXON_DT_MAX_HOURS * 3600.0
 	temporal = max(0.0, 1.0 - (abs(delta_seconds) / dt_max)) if dt_max > 0 else 0.0
-	return round(cfg.AXON_ALPHA * similarity + (1.0 - cfg.AXON_ALPHA) * temporal, 4)
+	return float(round(cfg.AXON_ALPHA * similarity + (1.0 - cfg.AXON_ALPHA) * temporal, 4))
 
 
 def _split_associations(payload: dict, own_collection: str) -> Tuple[List[Any], List[Axon]]:
